requirements.txt
langchain
faiss-cpu
sentence-transformers
streamlit
langchain-google-genai
langchain-community
python-dotenv
pathlib